+++ src/components/add-edit-customer/customer-sales-channel.tsx
@@ -4,20 +4,21 @@ import TextInput from "../inputs/textInput";
 import LoadingButton from "../button";
 import { useAppDispatch, useAppSelector } from "../../stores/hooks";
 import LinkButton from "../button/linkButton";
-import {
-  clearAllCustomerInfo,
-  updateCustomerSalesChannel,
-} from "../../stores/inAppDataInterations/addEditCustomerInfo";
+import { updateCustomerSalesChannel } from "../../stores/inAppDataInterations/addEditCustomerInfo";
 import Switch from "../switch";
 import Select from "../inputs/select";
-import { openSnackbar } from "../../stores/appFunctionality/snackbar";
-import {
-  addCustomersToList,
-  replaceCustomersInList,
-} from "../../stores/apiData/customers-lists";
 import useAxiosMultipart from "../../useHooks/useAxiosMultipart";
+import { customerRequestPayload } from "../../types/apiData/customers/request-payload";
 
-export default function AddEditCustomerSalesChannel({ id }: { id?: string }) {
+export default function AddEditCustomerSalesChannel({
+  id,
+  handleSubmit,
+  isSubmitting,
+}: {
+  id?: string;
+  handleSubmit: (payload: customerRequestPayload) => void;
+  isSubmitting: boolean;
+}) {
   const axios = useAxiosMultipart();
   const dispatch = useAppDispatch();
   const navigate = useNavigate();
@@ -31,7 +32,6 @@ export default function AddEditCustomerSalesChannel({ id }: { id?: string }) {
   const [calculateCommission, setCalculateCommission] = useState("");
   const [applyCommission, setApplyCommission] = useState("");
 
-  const [isSubmitting, setIsSubmitting] = useState(false);
   // populate apartment details interface
   useEffect(() => {
     const {
@@ -53,7 +53,6 @@ export default function AddEditCustomerSalesChannel({ id }: { id?: string }) {
     async (e: SyntheticEvent) => {
       // console.log("triggered");
       e.preventDefault();
-      setIsSubmitting(true);
       const { customerDetails, customerVerification, customerCompany } =
         storeCustomerDatast;
       const payload = {
@@ -76,33 +75,8 @@ export default function AddEditCustomerSalesChannel({ id }: { id?: string }) {
         })
       );
       try {
-        if (id) {
-          const response = await axios.post(`/admin/user/${id}`, payload);
-          const data = response?.data?.data;
-          dispatch(
-            openSnackbar({
-              message: "Customer's information successfully updated",
-              isError: false,
-            })
-          );
-          dispatch(replaceCustomersInList(data));
-        } else {
-          const response = await axios.post(`/admin/user`, payload);
-          const data = response?.data?.data;
-          dispatch(
-            openSnackbar({
-              message: "Customer information successfully created",
-              isError: false,
-            })
-          );
-          dispatch(addCustomersToList(data));
-        }
-        dispatch(clearAllCustomerInfo());
-        navigate(`/customers`);
-      } catch (error) {
-      } finally {
-        setIsSubmitting(false);
-      }
+        handleSubmit(payload);
+      } catch (error) {}
     },
     [
       storeCustomerDatast,