+++ src/components/tables/mobile/roles.tsx
@@ -48,7 +48,7 @@ export default function MobileRolesTable({
                       </span>{" "}
                       <span className="z-10 text-center group-hover:flex hidden w-52 bg-white text-sm absolute right-0 top-0 rounded-lg shadow-lg flex-col">
                         <Link
-                          to={`/admin-roles/edit/${item?.id}`}
+                          to={`/admin/admin-roles/edit/${item?.id}`}
                           className="p-3 px-4 hover:bg-primary/10 transition-all rounded-lg"
                         >
                           Edit role