+++ src/services-hooks/useGetCustomerById.tsx
@@ -81,7 +81,7 @@ export default function useGetCustomerById(id?: string) {
           last_name,
           email,
           phone,
-          profile_photo: { name: "", size: 1000, preview: profile_photo },
+          profile_photo: { id: 1000, preview: profile_photo },
           gender: gender || "Male",
           dob: new Date(dob)?.toISOString()?.slice(0, 10),
           country: country,
@@ -95,7 +95,10 @@ export default function useGetCustomerById(id?: string) {
           place_of_birth: place_of_birth,
           id_type: id_type,
           id_number: id_number,
-          identity_document: { name: "", size: 0, preview: "" },
+          identity_document: {
+            id: 1001,
+            preview: identity_verification_document,
+          },
           password: "",
           notes: notes,
         })