+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+##### 06.01.2025
+
+> Fixed bugs, testing and inconsistencies resolutions
+
+##### 05.01.2025
+
+> Fixed Updating customer information bug
+
 ##### 04.01.2025
 
 > Identity document approval endpoint integration